Separating Sentry fixtures and task test
