+++ pytest.ini
@@ -15,6 +15,7 @@ markers =
     integration_bigquery
     integration_saas
     integration_mailchimp
+    integration_sentry
     unit_saas
 
 filterwarnings =
